feat: make help menu available in logged-in mode
