+++ components/Themes/mode-toggle.tsx
@@ -23,8 +23,8 @@ export function ModeToggle() {
           variant="secondary"
           className="transition-colors duration-200 hover:bg-gray-400 dark:hover:bg-gray-600 focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
         >
-          <SunIcon className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-          <MoonIcon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
+          <SunIcon className="h-5 w-5 rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
+          <MoonIcon className="absolute h-5 w-5 rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
           <span className="sr-only">Toggle theme</span>
         </Button>
       </DropdownMenuTrigger>
